refactor: musical 모듈을 performance로 리네임
