--- musical/src/main/java/com/truve/platform/musical/service/domain/entity/MusicalSchedule.java
@@ -1,40 +1,0 @@
-package com.truve.platform.musical.service.domain.entity;
-
-import java.time.LocalDateTime;
-
-import com.truve.platform.common.support.BaseEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "musical_schedules")
-public class MusicalSchedule extends BaseEntity {
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "musical_id", nullable = false)
-	private Musical musical;
-
-	@Column(nullable = false)
-	private LocalDateTime dateTime;
-
-	@Column(nullable = false)
-	private Boolean isAvailable;
-
-	@Builder
-	private MusicalSchedule(Musical musical, LocalDateTime dateTime, Boolean isAvailable) {
-		this.musical = musical;
-		this.dateTime = dateTime;
-		this.isAvailable = isAvailable;
-	}
-}
