+++ performance/src/main/java/com/truve/platform/performance/PerformanceApplication.java
@@ -1,0 +1,14 @@
+package com.truve.platform.performance;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class PerformanceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PerformanceApplication.class, args);
+    }
+}
